time layers for geojson working
