+++ src/components/apis/apiMilitaryZones.js
@@ -19,7 +19,9 @@ async function militaryApicall(){
                 }
             }
         }));
-        return features;
+        const geojson = {'type': 'FeatureCollection',
+                        'features': features};
+        return geojson;
     } catch (error) {
         console.error('Error fetching data school:', error);
         throw error;
