+++ src/components/apis/apiTempIDEAM.js
@@ -4,6 +4,7 @@ async function apiTempIdeam(){
         const response = await fetch(API);
         const data = await response.json();
         const features  = await Promise.all(data.map(async (element) => {
+            element['fechaobservacion'] = String(new Date(element.fechaobservacion).getFullYear())
             return {
                 type: 'Feature',
                 properties: element,
@@ -13,7 +14,9 @@ async function apiTempIdeam(){
                 }
             }
         }));
-        return features;
+        const geojson = {'type': 'FeatureCollection',
+                        'features': features};
+        return geojson;
     } catch (error) {
         console.error('Error fetching data ideam temp:', error);
         throw error;
